backend/requirements.txt
fastapi==0.104.*
uvicorn[standard]==0.24.*
pydantic-settings==2.1.*
torch==2.1.*
torchvision==0.16.*
pillow==10.1.*
numpy==1.26.*
pyyaml==6.0.*
python-multipart==0.0.*
aiofiles==23.2.*
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

